Fixed Terrain Mesh Collision
Note: with the attached MOP Generator V2 script, the maximum X size and Z size are 256 for whatever reason (anything higher than that will mess up the generation). The solution for this is to just increase the Transform Scale of the Mesh Generator Parent Object to obtain larger terrain maps.